@@ -1,28 +1,36 @@
-import React from "react";
+import { useState } from "react";
 
-import ToggleButton from "./components/common/ToggleButton";
-import logo from "./logo.svg";
+import reactLogo from "./assets/react.svg";
+
+import viteLogo from "/vite.svg";
+
 import "./App.css";
+import ToggleButton from "./components/common/ToggleButton";
 
 function App() {
+    const [count, setCount] = useState(0);
+
     return (
-        <div className="App">
-            <header className="App-header dark:bg-blue-900">
-                <img src={logo} className="App-logo" alt="logo" />
-                <p>
-                    Edit <code>src/App.tsx</code> and save to reload.
-                </p>
-                <a
-                    className="App-link"
-                    href="https://reactjs.org"
-                    target="_blank"
-                    rel="noopener noreferrer"
-                >
-                    Learn React
+        <>
+            <div className="dark:bg-blue-900">
+                <a href="https://vite.dev" target="_blank">
+                    <img src={viteLogo} className="logo" alt="Vite logo" />
                 </a>
-                <ToggleButton />
-            </header>
-        </div>
+                <a href="https://react.dev" target="_blank">
+                    <img src={reactLogo} className="logo react" alt="React logo" />
+                </a>
+            </div>
+            <h1>Vite + React</h1>
+            <div className="card">
+                <button onClick={() => setCount((count) => count + 1)}>count is {count}</button>
+                <p>
+                    Edit <code>src/App.tsx</code> and save to test HMR
+                </p>
+            </div>
+            <p className="read-the-docs">Click on the Vite and React logos to learn more</p>
+            <h1 className="text-3xl font-bold underline">Hello world!</h1>
+            <ToggleButton />
+        </>
     );
 }
 
